Refactor registry integration tests
